Loading and reformatting of metabolomics data

Read a data file, select columns necessary for analysis, and return the reformatted data.

Usage

preprocess(
  fileName,
  dataSet = NULL,
  names = c("gender", "treatment", "replicate")
)

Arguments

fileName

The name of the .csv file containing metabolomics data (including the path to the file, if needed).

dataSet

The raw data set, if already loaded in R.

names

A vector of strings (default = c("gender", "treatment", "replicate")) specifying the names of the attribute columns.

Value

A 2d dataframe.

Details

The function executes the following:

Reads the file.

Provides summary statistics and a histogram of all values reported in the data set.

Re-formats the data to present individual compounds as columns.

Stores the data as a data.frame and prints the levels of attributes to the user.